+++ Sources/VaporTesting/Docs.docc/index.md
@@ -1,0 +1,7 @@
+# ``VaporTesting``
+
+VaporTesting is a set of utilities to make writing unit tests for Vapor applications easy.
+
+## Overview
+
+For more details, check out the [testing documentation](https://docs.vapor.codes/advanced/testing/).
