+++ Sources/VaporTesting/Exports.swift
@@ -1,3 +1,2 @@
-@_documentation(visibility: internal) @_exported import XCTest
 @_documentation(visibility: internal) @_exported import Vapor
 @_documentation(visibility: internal) @_exported import VaporTestUtils
