+++ Sources/XCTVapor/+XCTHTTPResponse.swift
@@ -1,49 +1,4 @@
-import NIOCore
-import NIOHTTP1
-import NIOConcurrencyHelpers
-import Vapor
 import XCTest
-
-public struct XCTHTTPResponse: Sendable {
-    public var status: HTTPStatus
-    public var headers: HTTPHeaders
-    public var body: ByteBuffer
-}
-
-extension XCTHTTPResponse {
-    private struct _ContentContainer: ContentContainer {
-        var body: ByteBuffer
-        var headers: HTTPHeaders
-
-        var contentType: HTTPMediaType? {
-            return self.headers.contentType
-        }
-
-        mutating func encode<E>(_ encodable: E, using encoder: ContentEncoder) throws where E : Encodable {
-            fatalError("Encoding to test response is not supported")
-        }
-
-        func decode<D>(_ decodable: D.Type, using decoder: ContentDecoder) throws -> D where D : Decodable {
-            try decoder.decode(D.self, from: self.body, headers: self.headers)
-        }
-
-        func decode<C>(_ content: C.Type, using decoder: ContentDecoder) throws -> C where C : Content {
-            var decoded = try decoder.decode(C.self, from: self.body, headers: self.headers)
-            try decoded.afterDecode()
-            return decoded
-        }
-    }
-
-    public var content: ContentContainer {
-        _ContentContainer(body: self.body, headers: self.headers)
-    }
-}
-
-extension Response.Body {
-    var isEmpty: Bool {
-        return self.count == 0
-    }
-}
 
 public func XCTAssertContent<D>(
     _ type: D.Type,
@@ -85,7 +40,7 @@ public func XCTAssertContains(_ haystack: String?, _ needle: String?, file: Stat
 }
 
 public func XCTAssertEqualJSON<T>(_ data: String?, _ test: T, file: StaticString = #filePath, line: UInt = #line)
-    where T: Codable & Equatable
+where T: Codable & Equatable
 {
     guard let data = data else {
         XCTFail("nil does not equal \(test)", file: file, line: line)
@@ -98,4 +53,3 @@ public func XCTAssertEqualJSON<T>(_ data: String?, _ test: T, file: StaticString
         XCTFail("could not decode \(T.self): \(error)", file: (file), line: line)
     }
 }
-
