+++ Sources/XCTVapor/typealiases.swift
@@ -1,0 +1,3 @@
+public typealias XCTHTTPRequest = TestingHTTPRequest
+public typealias XCTHTTPResponse = TestingHTTPResponse
+public typealias XCTApplicationTester = TestingApplicationTester
